pyproject.toml
[project]
name = "reddit-rag"
version = "0.1.0"
description = "Reddit RAG Agent"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-qdrant>=0.2",
    "langgraph>=0.2",
    "qdrant-client>=1.12",
    "fastembed>=0.4",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
